Json Formatter: handle comments, fix the handling of tags.
Add the comments from the feature file to the Json output. Test cases
from Scenario Outlines have the comments from the Scenario Outline,
the Examples table and the Examples table row.
All tags applied to a test case are added to the tags of the test case,
tags on the Feature are added to all test cases, and test cases from
Scenario Outlines also have the tags for both the Scenario Outline and
the Examples table.